move util folder
