+++ util/helperFunctions.js
