+++ util/mininginfo.js
